SMRF: accept datagrams for local processing even if not coming from the preferred parent
Previous version was dropping a packet if it wasn't sent by the
preferred patent. For forwarding, this means avoiding duplicates in the
forwarding process. For local processing, it means avoiding duplicates
at the higher layer protocol.

This patch enables the latter, increasing the chances that the packet
gets through, at the price of possible duplicates delivered, but without
increase in network traffic.

Signed-off-by: Csaba Kiraly <kiraly@disi.unitn.it>